Rename extract_expected_assertions_errors.js -> extractExpectedAssertionsErrors.js
- Rename file & imports
- No test found
